requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
cloudinary==1.34.0
sendgrid==6.10.0
Pillow==10.0.0
flasgger==0.9.7.1
gunicorn==21.2.0